update: add handlers for report and end as consumer widgets
